+++ docs/Changelog.md
@@ -4,7 +4,7 @@
 - Added `mobileTokenData` to authorize request for passing customer-specific data ([documentation](Using-Operations.md#Passing-Additional-Mobile-Token-Data))
   - Available with PowerAuth server 1.10+
   - Can be used for fraud detection systems (FDS) or other custom business logic
-- `register` method in `WMTPush` now accepts `WMTPushPlatform` object instead of separate parameters
+- `register` method in `WMTPush` now accepts `WMTPushData` object instead of separate parameters
   - for more information, see [migration guide](Version-2.1.md)
 
 ## 2.0.0 (6/2025)